% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R, R/auc.coxph.R
\name{auc}
\alias{auc}
\alias{auc.cph}
\alias{auc.coxph}
\title{auc for model}
\usage{
auc(...)

\method{auc}{cph}(..., model = NULL, x = NULL, method = c("NNE", "KM"))

\method{auc}{coxph}(..., model = NULL, x = NULL, method = c("NNE", "KM"))
}
\arguments{
\item{...}{one or more fit}

\item{model}{can be logical or characters. FALSE means no model TP and FP,
characters mean model names.}

\item{x}{can be logical or characters. TRUE means all x variable in regression
will be calculated. One or more characters will be calculated only.}

\item{method}{NNE or KM}
}
\value{
auc dataframe

one auc_coxph for cox regression. model means model names,
}
\description{
auc for model
}
