% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\alias{summary.roc_coxph}
\alias{summary.roc_logit}
\alias{summary.auc_coxph}
\title{Extract data from roc() function}
\usage{
\method{summary}{roc_coxph}(x, incomparables = FALSE, ...)

\method{summary}{roc_logit}(x, incomparables = FALSE, ...)

\method{summary}{auc_coxph}(x, incomparables = FALSE, ...)
}
\arguments{
\item{x}{result of roc() function}

\item{incomparables}{ignore}

\item{...}{ignore}
}
\value{
a dataframe.
}
\description{
Extract data from roc() function
}
