% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique.R
\name{unique}
\alias{unique}
\alias{unique.roc_coxph}
\alias{unique.roc_logit}
\alias{unique.auc_coxph}
\title{Extract data from roc() function}
\usage{
\method{unique}{roc_coxph}(x, incomparables = FALSE, ...)

\method{unique}{roc_logit}(x, incomparables = FALSE, ...)

\method{unique}{auc_coxph}(x, incomparables = FALSE, ...)
}
\arguments{
\item{x}{result of roc() function}

\item{incomparables}{ignore}

\item{...}{ignore}
}
\value{
a dataframe.
}
\description{
Extract data from roc() function
}
